etl/requirements.txt
pyyaml==6.0.1
python-dotenv==1.0.1
schedule==1.2.2
sqlalchemy==2.0.30
requests==2.32.2
pandas==2.2.2
pg8000==1.31.2
